python/requirements.txt
ultralytics==8.0.114
pandas
opencv-python
numpy
scipy
easyocr
filterpy
flask
torch
pillow
